chat_vectores/requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
openai==0.28
sqlalchemy
python-dotenv
